Add files via upload
Initial pages created for our project, Rate My Class. Consist of a login (todo) on the project page, barebones home page, fetch_classes.php (finds classes when user searches for a department-- to implement), and submit_review.php (which populates a database with a review for a specific class-- yet to be completely implement).